requirements.txt
# NHANES Data Analysis Pipeline Requirements
# Python 3.8+

# Core data processing
pandas>=1.3.0
numpy>=1.21.0

# Web scraping and parsing
beautifulsoup4>=4.10.0
lxml>=4.6.0

# Logging
loguru>=0.6.0

# YAML processing
PyYAML>=6.0

# Optional: For development and testing
pytest>=7.0.0
pytest-cov>=3.0.0
black>=22.0.0
flake8>=4.0.0
mypy>=0.950
isort>=5.10.0

# Optional: For data visualization
matplotlib>=3.5.0
seaborn>=0.11.2

# Optional: For progress bars
tqdm>=4.62.0